Refactor to address review comments and move scan type decider changes to Autonomous package.
